+++ src/demo-single.js
@@ -6,7 +6,7 @@ const likertOptions = {
   responses: [
     { value: 1, text: 'Abysmal' },
     { value: 2, text: 'Poor' },
-    { value: 3, text: 'Average' },
+    { value: 3, text: 'Average', checked: true },
     { value: 4, text: 'Good' },
     { value: 5, text: 'Excellent' },
   ],
@@ -18,17 +18,24 @@ export default function DemoSingle() {
 
   return (
     <>
+      <h1>Interaction</h1>
       <Likert
         {...likertOptions}
         onChange={setChosen}
         ref={refLikert}
-        // onClick={() => console.log('pretzels')}
-        // id='steckerlfisch'
-        // className='donuts'
+        onClick={() => console.log('it was clicked')}
+        className='withBorder'
       />
       <p>
         You chose: <strong>{chosen.value}—{chosen.text}</strong>
       </p>
+      <style>{`
+        .withBorder {
+          padding: 1em;
+          border: thin solid darkgreen;
+          border-radius: 1em;
+        }
+      `}</style>
     </>
   );
 }