+++ CHANGELOG.md
@@ -6,7 +6,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 [upcoming]
+----------------------------
+
+  * Example formatting in a grid.
+
+### Added
+
+  * You can now have a radio button pre-checked on first render.
+
+### Changed
+  * The `question` prop is now optional so that you can format as you wish.
+  * The `id` props is now (mostly) required. If you have multiple Likert components on a page ...
+  
+
+[4.0.0-1 alpha] 2021-02-06
+----------------------------
 
 ### Added
 
@@ -26,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [3.0.2] - 2021-01-22
+----------------------------
 
 ### Added